Add Tag in SearchTransaction
