+++ app/src/main/java/com/money/manager/ex/common/MmxBaseFragmentActivity.java
@@ -34,7 +34,7 @@ import android.view.View.OnClickListener;
 import com.money.manager.ex.R;
 import com.money.manager.ex.core.Core;
 import com.money.manager.ex.core.UIHelper;
-import com.money.manager.ex.errorhandle.MyExceptionHandler;
+import com.money.manager.ex.errorhandle.CrashReporter;
 import com.money.manager.ex.log.ErrorRaisedEvent;
 import com.money.manager.ex.settings.AppSettings;
 
@@ -56,6 +56,8 @@ public abstract class MmxBaseFragmentActivity
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        new CrashReporter(this);
+
         setTheme();
 
         AppSettings settings = new AppSettings(this);
@@ -72,7 +74,6 @@ public abstract class MmxBaseFragmentActivity
         this.compositeSubscription = new CompositeSubscription();
 
         super.onCreate(savedInstanceState);
-        new MyExceptionHandler(MmxBaseFragmentActivity.this);
 
         // Initialize the ActivityResultLauncher
         openDocumentLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(),
