+++ app/src/main/res/layout/activity_auth.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.money.manager.ex.errorhandle.AuthActivity">
+    tools:context="com.money.manager.ex.errorhandle.CrashReportActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -73,7 +73,7 @@
             android:paddingRight="@dimen/activity_horizontal_margin"
             android:paddingTop="@dimen/activity_vertical_margin"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
-            tools:context="com.money.manager.ex.errorhandle.AuthActivity"
+            tools:context="com.money.manager.ex.errorhandle.CrashReportActivity"
             tools:showIn="@layout/activity_auth">
 
             <EditText
